CHECKPOINT_1/requirements.txt
laspy                     2.5.4           py313h249b317_1    conda-forge
openai                    1.82.1             pyhd8ed1ab_0    conda-forge
python-dotenv             1.1.0              pyh29332c3_1    conda-forge
rasterio                  1.4.3           py313h9a45b90_1    conda-forge
